+++ main.js
@@ -2,61 +2,113 @@
 /* global define, brackets, $, require, Mustache */
 
 /*
-  HTML Skeleton
-  Created 2014 Triangle717
-  <http://Triangle717.WordPress.com/>
-
-  Licensed under The MIT License
-  <http://opensource.org/licenses/MIT/>
-*/
-
-
-/* ------- Begin Module Importing ------- */
+ * HTML Skeleton
+ * Created 2014 Triangle717
+ * <http://Triangle717.WordPress.com/>
+ *
+ * Licensed under The MIT License
+ * <http://opensource.org/licenses/MIT/>
+ */
 
 
 define(function (require, exports, module) {
   "use strict";
 
   // Import the required Brackets modules
-  var AppInit         = brackets.getModule("utils/AppInit"),
-      CommandManager  = brackets.getModule("command/CommandManager"),
-      Dialogs         = brackets.getModule("widgets/Dialogs"),
-      Document        = brackets.getModule("document/Document"),
-      EditorManager   = brackets.getModule("editor/EditorManager"),
-      ExtensionUtils  = brackets.getModule("utils/ExtensionUtils"),
-      ImageViewer     = brackets.getModule("editor/ImageViewer"),
-      Menus           = brackets.getModule("command/Menus"),
-      ProjectManager  = brackets.getModule("project/ProjectManager"),
-
-      // Import dialog localization
-      Strings         = require("strings"),
-
-      // Pull in the extension dialog
-      skellyDialogHtml    = require("text!htmlContent/mainDialog.html"),
-
+  var AppInit            = brackets.getModule("utils/AppInit"),
+      CommandManager     = brackets.getModule("command/CommandManager"),
+      Dialogs            = brackets.getModule("widgets/Dialogs"),
+      EditorManager      = brackets.getModule("editor/EditorManager"),
+      ExtensionUtils     = brackets.getModule("utils/ExtensionUtils"),
+      FileSystem         = brackets.getModule("filesystem/FileSystem"),
+      FileUtils          = brackets.getModule("file/FileUtils"),
+      ImageViewer        = brackets.getModule("editor/ImageViewer"),
+      LanguageManager    = brackets.getModule("language/LanguageManager"),
+      Menus              = brackets.getModule("command/Menus"),
+      PreferencesManager = brackets.getModule("preferences/PreferencesManager"),
+      ProjectManager     = brackets.getModule("project/ProjectManager"),
+      Strings            = require("strings"),
+      skeletonDialogHtml = require("text!htmlContent/mainDialog.html"),
+      toolbarButtonCode  = '<a href="#" id="html-skeleton-toolbar">',
       // Grab the logo to display in the dialog
-      skellyLogo      = require.toUrl("img/HTML-Skeleton.svg"),
-      EXTENSION_ID    = "le717.html-skeleton";
+      skeletonLogo       = require.toUrl("img/HTML-Skeleton.svg"),
+      EXTENSION_ID       = "le717.html-skeleton",
+      // User's indent settings
+      indentUnits        = "",
+      // Localize the dialog box
+      localizedDialog    = Mustache.render(skeletonDialogHtml, Strings),
+      // Valid image files + SVG (as supported by Brackets)
+      imageFiles         = LanguageManager.getLanguage("image")._fileExtensions.concat("svg");
 
 
-  /* ------- End Module Importing ------- */
+  /* ------- Begin Polyfills ------- */
+
+
+  /**
+   * @private
+   * Polyfill, taken from http://stackoverflow.com/a/4550005
+   */
+  function _repeat(str, num) {
+    return (new Array(num + 1)).join(str);
+  }
+
+
+  /* ------- End Polyfills ------- */
+
+
+  /* ------- Begin Reading Indentation Preference ------- */
+
+
+  /**
+   * @private
+   * Get the user's indentation settings for inserted code
+   */
+  function _getIndentSize() {
+    var newIndentUnits, indentUnitsInt,
+        tabCharPref  = PreferencesManager.get("useTabChar", PreferencesManager.CURRENT_PROJECT);
+
+    // The user is using tabs
+    if (tabCharPref) {
+      indentUnitsInt = PreferencesManager.get("tabSize");
+      newIndentUnits = _repeat("\u0009", indentUnitsInt);
+
+      // The user is using spaces
+    } else {
+      indentUnitsInt = PreferencesManager.get("spaceUnits");
+      newIndentUnits = _repeat("\u0020", indentUnitsInt);
+    }
+    return newIndentUnits;
+  }
+
+  // Get user's indentation settings
+  PreferencesManager.on("change", function (e, data) {
+    data.ids.forEach(function (value, index) {
+
+      // A relevant preference was changed, update our settings
+      if (value === "useTabChar" || value === "tabSize" || value === "spaceUnits") {
+        // Do NOT attempt to assign `indentUnits` directly to the function.
+        // It will completely break otherwise.
+        var tempVar  = _getIndentSize();
+        indentUnits  = tempVar;
+      }
+    });
+  });
+
+
+  /* ------- End Reading Indentation Preference ------- */
 
 
   /* ------- Begin Available HTML Elements ------- */
 
-  // Assign a variable for 2 space indentation for easier coding
-  // FUTURE Replace with Sprint 37 preferences system
-  var twoSpaceIndent = "\u0020\u0020";
 
   // Placeholder variables for image size
-  var $imgWidth = 0,
+  var $imgWidth  = 0,
       $imgHeight = 0;
 
-  var skellyBones = [
+  var skeletonBones = [
     // Only the head and body tags + title and meta
-    '<!DOCTYPE html>\n<html lang="">\n<head>\n' + twoSpaceIndent +
-    '<meta charset="UTF-8">\n' + twoSpaceIndent + '<title></title>\n' +
-    '\n</head>\n\n<body>\n' + twoSpaceIndent + '\n</body>\n</html>\n',
+    '<!DOCTYPE html>\n<html lang="">\n<head>\nindent-size<meta charset="UTF-8">\n' +
+    'indent-size<title></title>\n</head>\n\n<body>\nindent-size\n</body>\n</html>\n',
 
     // External stylesheet
     '<link rel="stylesheet" href="">',
@@ -64,21 +116,17 @@ define(function (require, exports, module) {
     // Inline stylesheet
     '<style></style>',
 
-    // External script
+    // External (and edited to be inline) script
     '<script src=""></script>',
 
-    // Inline script
-    '<script></script>',
-
     // Full HTML skeleton
-    '<!DOCTYPE html>\n<html lang="">\n<head>\n' + twoSpaceIndent +
-    '<meta charset="UTF-8">\n' + twoSpaceIndent + '<title></title>\n' +
-    twoSpaceIndent + '<link rel="stylesheet" href="">' + '\n</head>\n\n<body>\n' +
-    twoSpaceIndent + '<script src=""></script>\n</body>\n</html>\n'
+    '<!DOCTYPE html>\n<html lang="">\n<head>\nindent-size<meta charset="UTF-8">\n' +
+    'indent-size<title></title>\nindent-size<link rel="stylesheet" href="">\n' +
+    '</head>\n\n<body>\nindent-size<script src=""></script>\n</body>\n</html>\n'
   ];
 
-  // Picture/Image
-  var imageCode = '<img src="" alt="" width="size-x" height="size-y" />';
+  // Image
+  var imageCode = '<img src="src-url" alt="" width="size-x" height="size-y">';
 
 
   /* ------- End Available HTML Elements ------- */
@@ -86,19 +134,26 @@ define(function (require, exports, module) {
 
   /* ------- Begin HTML Element Adding ------- */
 
-  function _insertAllTheCodes(finalElements) {
-    /* Inter the selected elements into the document */
 
-    // Get the last active editor because the dialog steals focus
+  /**
+   * @private
+   * Insert the selected elements into the document
+   */
+  function _insertAllTheCodes(finalElements) {
+    // Get the last active editor
     var editor = EditorManager.getActiveEditor();
     if (editor) {
       // Get the cursor position
       var cursor = editor.getCursorPos();
 
       // Get the elements from the list in reverse so everything is added in the proper order
-      finalElements.reverse().forEach(function (value) {
+      finalElements.reverse().forEach(function (value, index) {
         //  Wrap the actions in a `batchOperation` call, per guidelines
         editor.document.batchOperation(function() {
+
+          // Do a regex search for the `indent-size` keyword
+          // and replace it with the user's indent settings
+          value = value.replace(/indent-size/g, indentUnits);
 
           // Insert the selected elements at the current cursor position
           editor.document.replaceRange(value, cursor);
@@ -114,54 +169,76 @@ define(function (require, exports, module) {
   /* ------- Begin HTML Element Choices ------- */
 
 
+  /**
+   * @private
+   * Get element choices
+   */
   function _getOptions() {
-    /* Get element choices */
-
     var imageCodeNew,
         // Stores the elements to be added
         finalElements = [],
-
         // Store all the option IDs for easier access
-        optionIDs = ["#head-body", "#extern-style-tag", "#inline-style-tag",
-                     "#extern-script-tag", "#inline-script-tag", "#full-skelly"
-                    ],
-
+        optionIDs     = ["#head-body", "#extern-style-tag", "#inline-style-tag",
+                         "#extern-script-tag", "#inline-script-tag", "#full-skeleton"
+                        ],
         // Shortcuts to the image size input boxes
-        $imgWidthID = $("#img-width"),
-        $imgHeightID = $("#img-height");
+        $imgWidthID   = $(".html-skeleton #img-width"),
+        $imgHeightID  = $(".html-skeleton #img-height");
 
     // For each option that is checked, add the corresponding element
     // to `finalElements` for addition in document
-    optionIDs.forEach(function (value, index) {
-      if ($(value + ":checked").val() === "on") {
-        finalElements.push(skellyBones[index]);
+    optionIDs.forEach(function(value, index) {
+      if ($(".html-skeleton " + value).prop("checked")) {
+
+        // The inline script box was checked, reuse external script string
+        if (index === 4) {
+          finalElements.push(skeletonBones[3].replace(/\ssrc="">/, ">"));
+
+          // Because of the script element editing above, redirect the
+          // Full HTML Skeleton option to the proper index
+        } else if (index === 5) {
+          finalElements.push(skeletonBones[4]);
+
+        } else {
+          // It was another element that does not require editing/redirecting
+          finalElements.push(skeletonBones[index]);
+        }
       }
     });
 
-    // The picture/image box is checked
-    if ($("#img-tag:checked").val() === "on") {
+    // The picture/image box was checked
+    if ($(".html-skeleton #img-tag").prop("checked")) {
+      var $inputWidth  = $imgWidthID.val(),
+          $inputHeight = $imgHeightID.val();
 
-      // The width box was filled out, use that value
-      if ($imgWidthID.val()) {
-        $imgWidth = $imgWidthID.val();
-
-      } else {
-        // The width box was empty, reset to 0
-        $imgWidth = 0;
+      // The values could not be picked out,
+      // Use 0 instead
+      // NOTE: For SVGs, temp action until #15 is complete
+      switch ($inputWidth) {
+        case "":
+          $imgWidth = 0;
+          break;
+        // The width box was filled out, use that value
+        default:
+          $imgWidth = $inputWidth;
       }
 
-      // The height box was filled out, use that value
-      if ($imgHeightID.val()) {
-        $imgHeight = $imgHeightID.val();
-
-      } else {
-        // The height box was empty, reset to 0
-        $imgHeight = 0;
+      // The values could not be picked out,
+      // Use 0 instead
+      // NOTE: For SVGs, temp action until #15 is complete
+      switch ($inputHeight) {
+        case "":
+          $imgHeight = 0;
+          break;
+        // The height box was filled out, use that value
+        default:
+          $imgHeight = $inputHeight;
       }
 
       // Add the image tag to `finalElements` for addition in document,
       // replacing the invalid values with valid ones
-      imageCodeNew = imageCode.replace(/size-x/, $imgWidth);
+      imageCodeNew = imageCode.replace(/src-url/, $(".html-skeleton-image #img-src").text());
+      imageCodeNew = imageCodeNew.replace(/size-x/, $imgWidth);
       imageCodeNew = imageCodeNew.replace(/size-y/, $imgHeight);
       finalElements.push(imageCodeNew);
     }
@@ -174,44 +251,185 @@ define(function (require, exports, module) {
   /* ------- End HTML Element Choices ------- */
 
 
-  /* ------- Begin HTML Skeleton Dialog Box ------- */
+  /* ------- Begin HTML Skeleton Dialog Boxes ------- */
 
+  /**
+   * @private
+   * Display dialog box
+   */
+  function _handleSkeletonButton() {
+    var skeletonDialog = Dialogs.showModalDialogUsingTemplate(localizedDialog),
+        $dialog = skeletonDialog.getElement(),
+        $doneButton = $('.dialog-button[data-button-id="ok"]', $dialog);
 
-  function _showSkellyDialog() {
-    /* Display the HTML Skeleton box */
+    // If the Browse button is clicked, proceed to open the browse dialog
+    $('.dialog-button[data-button-id="browse"]', $dialog).on("click", function(e) {
+      _showImageFileDialog(e);
+    });
 
-    var localized = Mustache.render(skellyDialogHtml, Strings);
-    var skellyDialog = Dialogs.showModalDialogUsingTemplate(localized),
-        $doneButton = skellyDialog.getElement().find('.dialog-button[data-button-id="ok"]');
-
-    // Display logo using Bracket's viewer
-    ImageViewer.render(skellyLogo, $(".html-skeleton-image"));
+    // Display logo (and any user images) using Brackets' ImageViewer
+    ImageViewer.render(skeletonLogo, $(".html-skeleton-image"));
 
     // Hide image stats
-    $("#img-tip").remove();
-    $("#img-scale").remove();
+    $(".html-skeleton-image #img-tip").remove();
+    $(".html-skeleton-image #img-scale").remove();
 
     // Upon closing the dialog, run function to gather and apply choices
     $doneButton.on("click", _getOptions);
   }
 
+  /**
+   * @private
+   * Open the file browse dialog for the user to select an image
+   */
+  function _showImageFileDialog(e) {
+    // Only display the image if the user selects ones
+    FileSystem.showOpenDialog(false, false, Strings.FILE_DIALOG_TITLE, null, imageFiles,
+                              function (closedDialog, selectedFile) {
+                                if (!closedDialog && selectedFile && selectedFile.length > 0) {
+                                  _handleImage(selectedFile[0]);
+                                }
+                              });
+    e.preventDefault();
+    e.stopPropagation();
+  }
 
-  /* ------- End HTML Skeleton Dialog Box ------- */
+
+  /* ------- End HTML Skeleton Dialog Boxes ------- */
+
+
+  /* ------- Begin user-selected image display ------- */
+
+  /**
+   * @private
+   * Various image path utilities
+   */
+  function _imgPathUtils(imgPath) {
+    // Make the image path relative (if possible)
+    imgPath = ProjectManager.makeProjectRelativeIfPossible(imgPath);
+
+    // If the path is longer than 50 characters, split it up for better displaying
+    if (imgPath.length > 50) {
+      imgPath = imgPath.substring(0, 51) + "<br>" + imgPath.substring(51, imgPath.length);
+    }
+    return imgPath;
+  }
+
+  /**
+   * @private
+   * Display the user selected image
+   */
+  function _handleImage(userImageFile) {
+    var imageWidth     = 0,
+        imageHeight    = 0,
+        supportedImage = true,
+        $imgWidth      = $(".html-skeleton #img-width"),
+        $imgHeight     = $(".html-skeleton #img-height"),
+        $imgCheckBox   = $(".html-skeleton #img-tag"),
+        $imgPreview    = $(".html-skeleton-image #img-preview"),
+        $showImgPath   = $(".html-skeleton-image #img-src"),
+        $imgErrorText  = $(".html-skeleton-image #img-error-text");
+
+    if (brackets.platform !== "mac") {
+      // Assume otherwise on other platforms as the file filter drop down is ignored but on Mac (https://trello.com/c/430aXkpq)
+      supportedImage = false;
+
+      // Go through the supported image list and check if the image is supported
+      imageFiles.forEach(function(value, index) {
+        if (FileUtils.getFileExtension(userImageFile) === value) {
+          // Yes, the image is supported
+          supportedImage = true;
+        }
+      });
+    }
+
+    // The Image check box was not checked before now. Since the user has opened an image,
+    // let's assume the user wants to use it and check the box.
+    if (!$imgCheckBox.prop("checked")) {
+      $imgCheckBox.prop("checked", true);
+    }
+
+    // The image is not a supported file type
+    if (!supportedImage) {
+      // Reset the width and height fields
+      $imgWidth.val("");
+      $imgHeight.val("");
+
+      // Run process to trim the path
+      userImageFile = _imgPathUtils(userImageFile);
+
+      // Update display for image
+      $showImgPath.html("");
+      $showImgPath.html(userImageFile);
+      $imgErrorText.html("<br>is not supported for previewing!");
+      $showImgPath.css("color", "red");
+      $imgPreview.removeClass("html-skeleton-image-shadow");
+
+      // Display extension logo
+      $imgPreview.attr("src", skeletonLogo);
+      return false;
+
+      // The image is a supported file type, move on
+    } else {
+      // Display the image using the full path
+      $imgPreview.attr("src", userImageFile);
+
+      // Clean possible CSS applied from previewing an invalid image
+      $imgErrorText.html("");
+      $showImgPath.css("color", "");
+
+      // Position and add small shadow to container
+      $(".html-skeleton-image").css("position", "relative");
+      $imgPreview.addClass("html-skeleton-image-shadow");
+
+      // Run process to trim the path
+      userImageFile = _imgPathUtils(userImageFile);
+
+      // Show the file path
+      $showImgPath.html("");
+      $showImgPath.html(userImageFile);
+
+      // Get the image width and height
+      $imgPreview.bind("load", function() {
+        imageWidth  = $imgPreview[0].naturalWidth;
+        imageHeight = $imgPreview[0].naturalHeight;
+
+        // If the image width and heights are not zero, update the size inputs with the values
+        if (imageWidth) {
+          $imgWidth.val(imageWidth);
+        }
+        if (imageHeight) {
+          $imgHeight.val(imageHeight);
+        }
+      });
+      return true;
+    }
+  }
+
+
+  /* ------- End user-selected image display ------- */
 
 
   /* ------- Begin Extension Initialization ------- */
 
-
-  AppInit.appReady(function () {
-    /* Load the extension after Brackets itself has finished loading */
-
+  /**
+   * @private
+   * Load the extension after Brackets itself has finished loading
+   */
+  AppInit.appReady(function() {
     // Load extension CSS
     ExtensionUtils.loadStyleSheet(module, "css/style.css");
 
     // Create a menu item in the Edit menu
-    CommandManager.register(Strings.INSERT_HTML_ELEMENTS, EXTENSION_ID, _showSkellyDialog);
+    CommandManager.register(Strings.INSERT_HTML_ELEMENTS, EXTENSION_ID, _handleSkeletonButton);
     var menu = Menus.getMenu(Menus.AppMenuBar.EDIT_MENU);
     menu.addMenuItem(EXTENSION_ID);
+
+    // Create toolbar icon
+    var $toolbarButton = $(toolbarButtonCode);
+    $toolbarButton.appendTo("#main-toolbar > .buttons");
+    $toolbarButton.attr("title", Strings.DIALOG_TITLE);
+    $toolbarButton.on("click", _handleSkeletonButton);
   });
 });